+++ SylixOS/system/device/spipe/spipeLib.c
@@ -833,7 +833,7 @@ INT  _SpipeIoctl (PLW_SPIPE_FILE pspipefil,
     
     case FIOWAITABORT:                                                  /*  停止当前等待 IO 线程        */
         LW_SPIPE_LOCK(pspipedev, return (PX_ERROR));
-        if ((INT)piArgPtr & OPT_RABORT) {
+        if ((INT)(LONG)piArgPtr & OPT_RABORT) {
             ULONG  ulBlockNum;
             API_SemaphoreBStatus(pspipedev->SPIPEDEV_hReadLock, LW_NULL, LW_NULL, &ulBlockNum);
             if (ulBlockNum) {
@@ -841,7 +841,7 @@ INT  _SpipeIoctl (PLW_SPIPE_FILE pspipefil,
                 API_SemaphoreBPost(pspipedev->SPIPEDEV_hReadLock);      /*  激活读等待线程              */
             }
         }
-        if ((INT)piArgPtr & OPT_WABORT) {
+        if ((INT)(LONG)piArgPtr & OPT_WABORT) {
             ULONG  ulBlockNum;
             API_SemaphoreBStatus(pspipedev->SPIPEDEV_hWriteLock, LW_NULL, LW_NULL, &ulBlockNum);
             if (ulBlockNum) {
@@ -860,7 +860,7 @@ INT  _SpipeIoctl (PLW_SPIPE_FILE pspipefil,
         
     case FIOPIPENOSIG:                                                  /*  不需要信号                  */
         LW_SPIPE_LOCK(pspipedev, return (PX_ERROR));
-        if ((INT)piArgPtr) {
+        if ((INT)(LONG)piArgPtr) {
             pspipefil->SPIPEFIL_iExtMode |= LW_SPIPE_EXT_MODE_NOSIG;
         } else {
             pspipefil->SPIPEFIL_iExtMode &= ~LW_SPIPE_EXT_MODE_NOSIG;
